+++ examples/with-i18n/src/locales.ts
@@ -1,0 +1,8 @@
+export const messages: Record<string, any> = {
+  'en-US': {
+    buttonText: 'Normal Button',
+  },
+  'zh-CN': {
+    buttonText: '普通按钮',
+  },
+};