+++ packages/ice/tests/fixtures/removeCode/function-exports.ts
@@ -1,3 +1,0 @@
-export default function Bar() {}
-export function pageConfig() {}
-export function getData() {}