+++ packages/plugin-i18n/src/utils/removeBasenameFromPath.ts
@@ -1,0 +1,18 @@
+export default function removeBasenameFromPath(pathname: string, basename?: string) {
+  if (typeof basename !== 'string') {
+    return pathname;
+  }
+  if (basename[0] !== '/') {
+    // compatible with no slash. For example: ice -> /ice
+    basename = `/${basename}`;
+  }
+
+  if (pathname.startsWith(basename)) {
+    pathname = pathname.substring(basename.length);
+    if (!pathname.startsWith('/')) {
+      pathname = `/${pathname || ''}`;
+    }
+  }
+
+  return pathname;
+}