+++ packages/plugin-rax-compat/src/index.ts
@@ -57,6 +57,7 @@ let warnOnce = false;
 
 export interface CompatRaxOptions {
   inlineStyle?: boolean;
+  cssModule?: boolean;
 }
 
 const plugin: Plugin<CompatRaxOptions> = (options = {}) => ({
@@ -114,12 +115,14 @@ const plugin: Plugin<CompatRaxOptions> = (options = {}) => ({
           warnOnce = true;
         }
 
+        const transformCssModule = options.cssModule == null ? true : options.cssModule;
+
         if (userConfig.ssr || userConfig.ssg) {
-          config.server.buildOptions = applyStylesheetLoaderForServer(config.server.buildOptions);
+          config.server.buildOptions = applyStylesheetLoaderForServer(config.server.buildOptions, transformCssModule);
         }
 
         config.configureWebpack ??= [];
-        config.configureWebpack.unshift(styleSheetLoaderForClient);
+        config.configureWebpack.unshift((config) => styleSheetLoaderForClient(config, transformCssModule));
         config.transforms = [
           ...(config.transforms || []),
           getClassNameToStyleTransformer(userConfig.syntaxFeatures || {}),
@@ -140,6 +143,7 @@ function getClassNameToStyleTransformer(syntaxFeatures) {
   const plugins: (string | Array<string | object>)[] = [
     [require.resolve('babel-plugin-transform-jsx-stylesheet'), {
       retainClassName: true,
+      forceEnableCSS: true,
     }],
   ];
 
@@ -198,16 +202,18 @@ function getClassNameToStyleTransformer(syntaxFeatures) {
  * StyleSheet Loader for CSR.
  * Transform css files to inline style by webpack loader.
  */
-const styleSheetLoaderForClient = (config) => {
+const styleSheetLoaderForClient = (config, transformCssModule) => {
   const { rules } = config.module || {};
   if (Array.isArray(rules)) {
     for (let i = 0, l = rules.length; i < l; i++) {
       const rule: RuleSetRule | any = rules[i];
       // Find the css rule, that default to CSS Modules.
       if (rule.test && rule.test instanceof RegExp && rule.test.source.indexOf('.css') > -1) {
+        rule.test = transformCssModule ? /(\.module|global)\.css$/i : /(\.global)\.css$/i;
         rules[i] = {
           test: /\.css$/i,
           oneOf: [
+            rule,
             ruleSetStylesheet,
           ],
         };
@@ -215,9 +221,11 @@ const styleSheetLoaderForClient = (config) => {
 
       // Find and replace the less rule
       if (rule.test && rule.test instanceof RegExp && rule.test.source.indexOf('.less') > -1) {
+        rule.test = transformCssModule ? /(\.module|global)\.css$/i : /(\.global)\.css$/i;
         rules[i] = {
           test: /\.less$/i,
           oneOf: [
+            rule,
             ruleSetStylesheetForLess,
           ],
         };
@@ -231,16 +239,16 @@ const styleSheetLoaderForClient = (config) => {
  * StyleSheet Loader for Server.
  * @param config
  */
-function applyStylesheetLoaderForServer(preBuildOptions) {
+function applyStylesheetLoaderForServer(preBuildOptions, transformCssModule) {
   return (buildOptions) => {
     const currentOptions = preBuildOptions?.(buildOptions) || buildOptions;
 
     // Remove esbuild-empty-css while use inline style.
     currentOptions.plugins = currentOptions.plugins?.filter(({ name }) => name !== 'esbuild-empty-css');
-    const cssModuleIndex = currentOptions.plugins?.findIndex(({ name }) => name === 'esbuild-css-modules');
+    const cssModuleIndex = currentOptions.plugins?.findIndex(({ name }) => name === 'esbuild-css-modules') as number;
 
     // Add custom transform for server compile.
-    currentOptions.plugins?.splice(cssModuleIndex as number, 0, inlineStylePlugin());
+    currentOptions.plugins?.splice(transformCssModule ? cssModuleIndex + 1 : cssModuleIndex, 0, inlineStylePlugin());
 
     currentOptions.treeShaking = true;
     return currentOptions;