+++ packages/runtime/src/runtime.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import * as ReactDOM from 'react-dom/client';
 import type { ComponentType } from 'react';
+import { routerHistory as history } from './history.js';
 import type {
   Renderer,
   AppContext,
@@ -15,6 +16,7 @@ import type {
   SetRender,
   AppRouterProps,
   ComponentWithChildren,
+  ResponseHandler,
 } from './types.js';
 import { useData, useConfig } from './RouteContext.js';
 import { useAppContext } from './AppContext.js';
@@ -32,6 +34,8 @@ class Runtime {
 
   private render: Renderer;
 
+  private responseHandlers: ResponseHandler[];
+
   public constructor(appContext: AppContext, runtimeOptions?: Record<string, any>) {
     this.AppProvider = [];
     this.appContext = appContext;
@@ -42,9 +46,16 @@ class Runtime {
     };
     this.RouteWrappers = [];
     this.runtimeOptions = runtimeOptions;
+    this.responseHandlers = [];
+    this.getAppRouter = this.getAppRouter.bind(this);
   }
 
-  public getAppContext = () => this.appContext;
+  public getAppContext = () => {
+    return {
+      ...this.appContext,
+      RouteWrappers: this.RouteWrappers,
+    };
+  };
 
   public setAppContext = (appContext: AppContext) => {
     this.appContext = appContext;
@@ -54,13 +65,17 @@ class Runtime {
     return this.render;
   };
 
-  public getAppRouter = () => this.AppRouter;
+  public getAppRouter<T>() {
+    return this.AppRouter as ComponentType<T>;
+  }
 
   public getWrappers = () => this.RouteWrappers;
 
   public async loadModule(module: RuntimePlugin | StaticRuntimePlugin | CommonJsRuntime) {
     let runtimeAPI: RuntimeAPI = {
       addProvider: this.addProvider,
+      addResponseHandler: this.addResponseHandler,
+      getResponseHandlers: this.getResponseHandlers,
       getAppRouter: this.getAppRouter,
       setRender: this.setRender,
       addWrapper: this.addWrapper,
@@ -69,6 +84,7 @@ class Runtime {
       useData,
       useConfig,
       useAppContext,
+      history,
     };
 
     const runtimeModule = ((module as CommonJsRuntime).default || module) as RuntimePlugin;
@@ -107,6 +123,14 @@ class Runtime {
 
   public setAppRouter: SetAppRouter = (AppRouter) => {
     this.AppRouter = AppRouter;
+  };
+
+  public addResponseHandler = (handler: ResponseHandler) => {
+    this.responseHandlers.push(handler);
+  };
+
+  public getResponseHandlers = () => {
+    return this.responseHandlers;
   };
 }